Uncommented user authentication (is a user Admin?)

+++ admin/deleteItem.php
@@ -13,24 +13,24 @@ require_once("../Gateways/UserGateway.php");
 session_start();
 
 //Check if user is Logged In
-//if (isset($_SESSION['ID']))
-//{
-//    //check if user is admin
-//    $userGateway = new UserGateway();
-//    $user = $userGateway->rowDataQueryByID($_SESSION['ID']);
-//    if ($user->isAdmin)
-//    {
+if (isset($_SESSION['ID']))
+{
+    //check if user is admin
+    $userGateway = new UserGateway();
+    $user = $userGateway->rowDataQueryByID($_SESSION['ID']);
+    if ($user->isAdmin)
+    {
         //Deletes Item
         deleteItem();
 
-//    } else
-//    {
-//        echo "Insufficient Privilege";
-//    }
-//} else
-//{
-//    echo '<p>Not Logged In</p><a href="http://webprog.cs.ship.edu/webprog25/account_management/Login.html">Login?</a>';
-//}
+    } else
+    {
+        echo "Insufficient Privilege";
+    }
+} else
+{
+    echo '<p>Not Logged In</p><a href="http://webprog.cs.ship.edu/webprog25/account_management/Login.html">Login?</a>';
+}
 
 
 //Deletes an Item from the DB
